bug fix: world holes
The world was saying it generated stuff but really it hadn't.
This was because spawn iterated through everything and
said that it was generated.

Also now generation code and instance building code is vastly simplified,
seeing as we're going to stick with either generating or instancing a single chunk
each frame during the loop.